Added Command: Inc Added Command: Dec Added Command: Buffer Added Command: Local
Fixed: Graphics errors now stops execution.

Changed: All expressions setting a variable will now try to set a local first if one with the same name exists
Changed: Jumpto, if a wait command or section has updat in name, endless looping will not be stopped.

+++ src/ximias/dk/au/cs/fh/Components/ArgManipulation.java
@@ -25,7 +25,7 @@ public class ArgManipulation {
     public static boolean isHex(String n){
         try{
             //noinspection ResultOfMethodCallIgnored
-            Integer.parseInt(n,16);
+            Long.parseLong(n,16);
         }catch (NumberFormatException e){
             return false;
         }
